XW-456 | Remove obsolete mixin
